[BUGFIX] Avoid implicit conversion to int in StringUtility
StringUtility::multibyteStringPad() does a happy division
feeding str_repeat() to evenly pad on left and right side.

The division result can be a float. With PHP 8.1,
str_repeat($string, float) emits E_DEPRECATED
"Implicit conversion from float looses precision.".

Avoid this with an explicit cast.

Resolves: #95772
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71970
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>

+++ typo3/sysext/core/Classes/Utility/StringUtility.php
@@ -203,7 +203,7 @@ class StringUtility
                 $padded = ((int)($leftPadCount / $pad_string_len)) * $pad_string_len;
                 $leftPad = str_repeat($pad_string, (int)($leftPadCount / $pad_string_len));
                 $leftPad .= mb_substr($pad_string, 0, $leftPadCount - $padded);
-                $string = $leftPad . $string . str_repeat($pad_string, ($length - $len)/$pad_string_len);
+                $string = $leftPad . $string . str_repeat($pad_string, (int)(($length - $len)/$pad_string_len));
                 $string .= mb_substr($pad_string, 0, ($length - $len) % $pad_string_len);
                 return $string;
         }
